Move Adapter into lewis.core
Just like we moved the device stuff into lewis.core,
it makes sense to do the same with the adapter base
class and supporting infrastructure.
